// lib/types/feed.types.ts
export interface FeedOptions {
    skipamount: number;
    pagesize: number;
    affinity_deep_search_min_value: number;
    affinity_deep_search_max_value: number;
    affinity_deep_talk_value: number;
    affinity_deep_like_value: number;
    affinity_deep_followage_value: number;
    ranking_media_effect: number;
    ranking_like_effect: number;
    ranking_time_effect: number;
    ranking_default_like_effect: number;
}